Refactor code into a Map object
Disable aura/named/overlap parts to enable a map of all cities.
These should be reenabled in the future using Map object to specify
which kind of map is required. It will need a part to forge list of
cities to be mapped with or without axis, links, aura, names and with a
color set.
Switch from config 'guildes' definition to local file parsing.

Signed-off-by: Olmagzar <olmagzar@gmail.com>