Refactor geometry details into a GameGeom class
Add ColorSet class and specific subclasses which define the getColorIdx
method to retrieve the index of the color defined in the given subclass.
Add argument parser to allow disabling creating the map of all players
and/or creating the map of activity status of the cities.
Use the file specified as database containing details of the players.

Stellar colorset class defines seven colors assigned according to the
estimated exploration area computed from the encounter score:
 - red: for cities which explored two full rings or less
 - orange: for cities which explored five full rings or less
 - yellow: for cities which explored eight full rings or less
 - white: for cities which explored eleven full rings or less
 - light blue: for cities which explored fourteen full rings or less
 - blue: for cities which explored seventeen full rings or less
 - dark blue: for cities which explored eighteen full rings or more
These colors were choosen to follow the stars color from red, the small
and cold ones, to blue the large and hot ones. It is not the exact
colorset of stars because the colors are too light and alike to be
correctly interpreted on the map. Adjustments were required.
This choice of colorset was inspired by a fellow member who noticed that
the first attempt to display the location of the cities in a graph
looked like a constellation.

Forest colorset class defines six colors assigned according to the
estimated period of inactivity of each city (some might be active but
not detected):
 - green: for cities active in the last six days
 - light green: for cities active in the last thirteen days
 - yellow: for cities active in the last tweenty days
 - red: for cities active in the last tweenty-six days
 - brown: for cities active in the last thirty-four days
 - black: for cities inactive since thirty-four days or more
These colors were chosen to give inactive cities a look of trees in the
fall or in the winter and to give active cities a look of trees in the
spring or in the summer.

usage: drawMap.py [-h] [--all | --no-all] [--actives | --no-actives]
                  filename

Signed-off-by: Olmagzar <olmagzar@gmail.com>